fix: delete account

+++ backend/authentication/views.py
@@ -263,8 +263,9 @@ class AccountDeleteView(generics.GenericAPIView):
         account = get_object_or_404(Account, id=id)
         user = get_object_or_404(User, account=account)
         user_data = model_to_dict(user)
-        address = get_object_or_404(Address, id=user_data.get('address'))
         user.delete()
-        address.delete()
+        if user_data.get('address'):
+            address = get_object_or_404(Address, id=user_data.get('address'))
+            address.delete()
         account.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
